+++ examples/create-vite/src/componants/Login.jsx
@@ -14,7 +14,7 @@ const Login = () => {
   /**
    * Login Handle, the callback function onClick from the "Login" button
    *
-   * This function demostrate a dummy authentication, using useSignIn function
+   * This function demostrate a fake authentication, using useSignIn function
    */
   const loginHandler = () => {
     // Assuming that, all network Request is successfull, and the user is authenticated
@@ -37,7 +37,7 @@ const Login = () => {
     )
   } else {
     // If not authenticated, use the login flow
-    // For Demostration, I'm using just a button to login.
+    // For Demonstration; I'm using just a button to log in.
     // In reality, there should be a form, validation, nwetowrk request and other things
     return (
       <button className='p-2 border rounded-md bg-cyan-500 hover:bg-cyan-700 text-white' onClick={loginHandler}>Log In!!</button>