+++ packages/next/src/NextAuth.tsx
@@ -31,7 +31,7 @@ interface NextAuthProps {
 
 /**
   *
-  * Next Js Page Wrapper component.
+  * Next.js Page Wrapper component.
   *
   * @param props - The Properties of the component
   * @returns - Compoent with React Auth Kit integrated
@@ -45,7 +45,7 @@ interface NextAuthProps {
   * }: {
   *   children: React.ReactNode
   * }) {
-  *   return <NextAuth fallbackPath={'/login'}>{children}</NextAuth>
+  *   return <NextAuth fallbackPath={"/login"}>{children}</NextAuth>
   * }
   * ```
   *
@@ -73,7 +73,7 @@ export default function NextAuth({fallbackPath, children}: NextAuthProps) {
     if (!isAuthenticated(context.value)) {
       // Redirect them to the /login page, but save the current location they
       // were trying to go to when they were redirected. This allows us to
-      // send them along to that page after they login, which is a nicer
+      // send them along to that page after they log in, which is a nicer
       // user experience than dropping them off on the home page.
       context.set(doSignOut());
       push(fp);