+++ packages/react-router/src/AuthOutlet.tsx
@@ -62,9 +62,9 @@ const AuthOutlet : React.FC<AuthOutletProps> = ({fallbackPath}) => {
   }
 
   if (!isAuthenticated()) {
-    // Redirect them to the /login page, but save the current location they
+    // Redirect them to the /login page but save the current location they
     // were trying to go to when they were redirected. This allows us to
-    // send them along to that page after they login, which is a nicer
+    // send them along to that page after they log in, which is a nicer
     // user experience than dropping them off on the home page.
     return <Navigate to={fp} replace />;
   }